Move files to their correct layer
