+++ app/src/main/java/com/example/exoplayersession/data/PipBroadcastReceiver.kt
@@ -1,0 +1,11 @@
+package com.example.exoplayersession.data
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class PipBroadcastReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context?, intent: Intent?) {
+        print("Try to pause the video")
+    }
+}
