+++ app/src/main/java/com/example/exoplayersession/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.exoplayersession
+package com.example.exoplayersession.ui
 
 import android.content.Intent
 import android.os.Bundle
